requirements.txt
azure-functions
http.client
google_trans_new